+++ src/animations/maximize-window-animation.h
@@ -28,7 +28,7 @@ class MaximizeWindowAnimation : public Animation {
   MaximizeWindowAnimation(const WindowSystem &windowSystem,
                           const WindowT &window, Color color,
                           Color borderColor);
-  void render(int percentage) override;
+  void render(double percentage) override;
 
  private:
   Rectangle maxSize;
